requirements.txt
fastapi
uvicorn
python-nmap
pymetasploit3
requests
pyyaml
requests